billing_system/requirements.txt
django
djangorestframework
psycopg2
celery
redis